+++ R/predicate.R
@@ -1,0 +1,9 @@
+
+pscalar <- function(x) {
+  length(x) == 1L
+}
+
+pstring <- function(x) {
+  is.character(x) && length(x) == 1L
+}
+