+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# renv 0.14.0 (UNRELEASED)
+# renv 0.13.1
 
 * `renv::clean()` gains the `actions` argument, allowing the caller to control
   which specific actions are taken during a `clean()` operation.